+++ dev/examples/draggable.tsx
@@ -13,5 +13,5 @@ const styleB = {
 }
 
 export const App = () => {
-    return <motion.div dragEnabled style={styleA} />
+    return <motion.div drag style={styleA} />
 }