+++ src/motion/utils/functionality.ts
@@ -93,7 +93,7 @@ export const getAnimateComponent = (
 }
 
 const gestureProps = [
-    "dragEnabled",
+    "drag",
     "onPan",
     "onPanStart",
     "onPanEnd",
@@ -110,7 +110,7 @@ export const isGesturesEnabled = (props: MotionProps) => {
     return gestureProps.some(key => props.hasOwnProperty(key))
 }
 
-export const isDragEnabled = (props: MotionProps) => !!props.dragEnabled
+export const isDragEnabled = (props: MotionProps) => !!props.drag
 
 export const Gestures = makeHookComponent(
     ({ innerRef, ...props }: AnimateProps) => useGestures(props, innerRef)